+++ packages/label-core/src/templates/index.ts
@@ -1,3 +1,4 @@
+export { labelFilename } from './filename'
 export { DEFAULT_GHS_STOCK, GHS_ELEMENTS, GHS_SIGNAL_WORDS, GHS_TYPE_DEFAULT } from './ghs'
 export type { GhsLabelData, GhsSignalWord, GhsSupplier } from './ghs'
 export { ANCHORS, ARTWORK_DEFAULT, anchorBox, panelFor } from './stock'